crypto_Based-sbox/checker/requirements.txt
pwntools
xoflib